MentorMind_AI-main/requirements.txt
fastapi
uvicorn[standard]
python-multipart
celery[redis]
redis
pydantic
requests
python-dotenv
celery==5.3.6
redis==5.0.1
opencv-python
python-dotenv
numpy
pydub
speechrecognition
transformers
torch
pillow
moviepy
onnxruntime
onnx
openai-whisper
gtts